fix(review): address code review findings for indexer health tracking
- Add download_id column to both SQLite and Postgres migrations so fresh
  installs have the column without relying solely on ensureSchemaIntegrity
- Remove unused grabbedIndexers sync.Map field from Server struct
- Add clarifying comment to pruneIndexerStats explaining it deletes recent
  (not old) data per the "Delete Last N hours" UI intent
- Fix ConfirmModal verification input to use fieldset/legend instead of
  the deprecated form-control/label DaisyUI pattern
- Replace inline SVG search icon with <Search /> from lucide-react
- Replace any type annotations in ProviderChart with ProviderStatus
- Decompose 1018-line IndexerHealth.tsx into focused sub-components:
  IndexerHealthChart, IndexerHealthSummary, IndexerHealthFilters,
  IndexerHealthCard, PruneStatsModal (each < 200 lines)
- Remove spurious blank lines and trailing whitespace
