+++ frontend/src/pages/HealthPage/components/ProviderHealth/ProviderChart.tsx
@@ -16,6 +16,7 @@ import {
 import { LoadingSpinner } from "../../../../components/ui/LoadingSpinner";
 import { usePoolMetrics, useProviderHistoricalStats } from "../../../../hooks/useApi";
 import { formatBytes } from "../../../../lib/utils";
+import type { ProviderStatus } from "../../../../types/api";
 
 const COLORS = ["#3b82f6", "#10b981", "#f59e0b", "#ef4444", "#8b5cf6", "#ec4899", "#06b6d4"];
 
@@ -80,7 +81,7 @@ export function ProviderChart() {
 		let total = 0;
 
 		if (poolData?.providers) {
-			poolData.providers.forEach((p: any) => {
+			poolData.providers.forEach((p: ProviderStatus) => {
 				pTotals[p.host] = 0;
 			});
 		}
@@ -99,7 +100,7 @@ export function ProviderChart() {
 				timeLabel = dateObj.toLocaleString(undefined, { month: "short", year: "2-digit" });
 			}
 
-			const provider = poolData?.providers?.find((p: any) => p.id === stat.provider_id || stat.provider_id.startsWith(p.host));
+			const provider = poolData?.providers?.find((p: ProviderStatus) => p.id === stat.provider_id || stat.provider_id.startsWith(p.host));
 			const normalizedID = provider ? provider.host : stat.provider_id.split(":")[0];
 
 			if (!groupedByTime[timeKey]) groupedByTime[timeKey] = { name: timeLabel };
